refactor: change monetary fields from Double to Long across order domain (BR-08)

+++ backend/src/main/java/com/foodya/foodya_backend/admin/controller/AdminOrderController.java
@@ -53,7 +53,7 @@ public class AdminOrderController {
   }
 
   @GetMapping("/metrics/revenue")
-  public ResponseEntity<Double> revenue(
+  public ResponseEntity<Long> revenue(
       @RequestParam UUID restaurantId,
       @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) Instant startDate,
       @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) Instant endDate
